Add failure status to database entity for failed jobs (#35)
This adds a failed status to an entity in the database if a job failed. For now, this happens in the job. However, if we want to use retries, we probably need to move that to a special exception handler.
This also adds the error_code and error_message to all EricaRequest classes and sets them if a job fails.

* Remove retries from job creation

* Change status to failed after failure in perform_job

* Use test database for test_erica_request_repository.py

* Add error_code and error_message to EricaRequest classes

* Add error_code and error_message to the entity if job failed